requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
djangorestframework==3.15.2
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
protobuf==5.29.3
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rich==13.9.4
setuptools==75.1.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.2
